# packages/openapi/openapi/mockos-management.v1.json
{
  "components": {
    "schemas": {
      "GetEnvironmentDiscoveryResponse200Definition1": {
        "anyOf": [
          {
            "type": "null"
          },
          {
            "type": "boolean"
          },
          {
            "type": "number"
          },
          {
            "type": "string"
          },
          {
            "items": {
              "$ref": "#/components/schemas/GetEnvironmentDiscoveryResponse200Definition1"
            },
            "type": "array"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/GetEnvironmentDiscoveryResponse200Definition1"
            },
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          }
        ]
      },
      "Problem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "minLength": 1,
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "requestId": {
            "minLength": 1,
            "type": "string"
          },
          "status": {
            "maximum": 599,
            "minimum": 400,
            "type": "integer"
          },
          "title": {
            "minLength": 1,
            "type": "string"
          },
          "type": {
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "requestId"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "accessKey": {
        "bearerFormat": "mockOS Access Key",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Authenticated self-hosted management operations currently implemented by the mockOS Worker.",
    "title": "mockOS management API",
    "version": "0.1.0"
  },
...
  "openapi": "3.1.0",
  "paths": {
    "/environments/{environmentId}": {
      "delete": {
        "description": "Permanently deletes an environment. MCP may omit environmentId to target the session cursor; a successful cursor-targeted delete clears that cursor, so retry with the deleted environmentId explicitly. HTTP always requires environmentId.",
        "operationId": "delete_environment",
        "parameters": [
          {
            "in": "path",
            "name": "environmentId",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 8,
              "pattern": "^[a-z0-9][a-z0-9_-]+$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The operation completed successfully."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A mockOS RFC 7807 problem response."
          }
        },
        "security": [
          {
            "accessKey": []
          }
        ],
        "summary": "Delete mock identity environment",
        "tags": [
          "Management"
        ],
        "x-mockos-effect": "destructive",
        "x-mockos-retry": "idempotent",
        "x-mockos-scopes": [
          "env:rw"
        ]
      },
      "put": {
        "description": "Updates mutable environment settings through MCP or configures the complete environment document through the self-hosted HTTP control route.",
        "operationId": "configure_environment",
        "parameters": [
          {
            "in": "path",
            "name": "environmentId",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 8,
              "pattern": "^[a-z0-9][a-z0-9_-]+$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "createdAt": {
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                    "type": "string"
                  },
                  "id": {
                    "maxLength": 64,
                    "minLength": 8,
                    "pattern": "^[a-z0-9][a-z0-9_-]+$",
                    "type": "string"
                  },
                  "idleTtlHours": {
                    "default": 168,
                    "maximum": 8760,
                    "minimum": 1,
                    "type": "integer"
                  },
                  "name": {
                    "maxLength": 80,
                    "minLength": 1,
                    "type": "string"
                  },
                  "provider": {
                    "enum": [
                      "entra",
                      "okta"
                    ],
                    "type": "string"
                  },
                  "requestLogLimit": {
                    "default": 10000,
                    "maximum": 100000,
                    "minimum": 100,
                    "type": "integer"
                  },
                  "seed": {
                    "maxLength": 256,
                    "minLength": 1,
                    "type": "string"
                  },
                  "tenantId": {
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "name",
                  "provider",
                  "seed",
                  "tenantId",
                  "createdAt"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "additionalProperties": false,
                      "properties": {
                        "createdAt": {
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                          "type": "string"
                        },
                        "id": {
                          "maxLength": 64,
                          "minLength": 8,
                          "pattern": "^[a-z0-9][a-z0-9_-]+$",
                          "type": "string"
                        },
                        "idleTtlHours": {
                          "default": 168,
                          "maximum": 8760,
                          "minimum": 1,
                          "type": "integer"
                        },
                        "name": {
                          "maxLength": 80,
                          "minLength": 1,
                          "type": "string"
                        },
                        "provider": {
                          "enum": [
                            "entra",
                            "okta"
                          ],
                          "type": "string"
                        },
                        "requestLogLimit": {
                          "default": 10000,
                          "maximum": 100000,
                          "minimum": 100,
                          "type": "integer"
                        },
                        "seed": {
                          "maxLength": 256,
                          "minLength": 1,
                          "type": "string"
                        },
                        "tenantId": {
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "provider",
                        "seed",
                        "tenantId",
                        "createdAt",
                        "idleTtlHours",
                        "requestLogLimit"
                      ],
                      "type": "object"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "properties": {
                        "requestId": {
                          "minLength": 1,
                          "type": "string"
                        }
                      },
                      "required": [
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "The operation completed successfully."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A mockOS RFC 7807 problem response."
          }
        },
        "security": [
          {
            "accessKey": []
          }
        ],
        "summary": "Configure mock identity environment",
        "tags": [
          "Management"
        ],
        "x-mockos-effect": "mutation",
        "x-mockos-retry": "idempotent",
        "x-mockos-scopes": [
          "env:rw"
        ]
      }
    },
...
    "/environments/{environmentId}/well-known": {
      "get": {
        "description": "Returns the provider-shaped OIDC discovery document produced for an explicit issuer base.",
        "operationId": "get_environment_discovery",
        "parameters": [
          {
            "in": "path",
            "name": "environmentId",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 8,
              "pattern": "^[a-z0-9][a-z0-9_-]+$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "issuer_base",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "additionalProperties": {
                        "$ref": "#/components/schemas/GetEnvironmentDiscoveryResponse200Definition1"
                      },
                      "properties": {
                        "authorization_endpoint": {
                          "format": "uri",
                          "type": "string"
                        },
                        "claims_supported": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "code_challenge_methods_supported": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "grant_types_supported": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "id_token_signing_alg_values_supported": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "issuer": {
                          "format": "uri",
                          "type": "string"
                        },
                        "jwks_uri": {
                          "format": "uri",
                          "type": "string"
                        },
                        "response_modes_supported": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "response_types_supported": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "scopes_supported": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "subject_types_supported": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "token_endpoint": {
                          "format": "uri",
                          "type": "string"
                        },
                        "token_endpoint_auth_methods_supported": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "userinfo_endpoint": {
                          "format": "uri",
                          "type": "string"
                        }
                      },
                      "required": [
                        "issuer",
                        "authorization_endpoint",
                        "token_endpoint",
                        "jwks_uri",
                        "response_types_supported",
                        "response_modes_supported",
                        "subject_types_supported",
                        "id_token_signing_alg_values_supported",
                        "scopes_supported",
                        "token_endpoint_auth_methods_supported",
                        "claims_supported",
                        "grant_types_supported",
                        "code_challenge_methods_supported"
                      ],
                      "type": "object"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "properties": {
                        "requestId": {
                          "minLength": 1,
                          "type": "string"
                        }
                      },
                      "required": [
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "The operation completed successfully."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A mockOS RFC 7807 problem response."
          }
        },
        "security": [
          {
            "accessKey": []
          }
        ],
        "summary": "Get environment OIDC discovery document",
        "tags": [
          "Management"
        ],
        "x-mockos-effect": "read",
        "x-mockos-retry": "safe",
        "x-mockos-scopes": [
          "env:ro"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "/__mockos/v1"
    }
  ],
...
}